Minor exploit package cleanup
