+++ internal/exploit/models.go
@@ -1,0 +1,20 @@
+package exploit
+
+type Output struct {
+	Name string
+	Team string
+	Out  []byte
+}
+
+type Result struct {
+	Flag    string `json:"flag"`
+	Exploit string `json:"sploit"`
+	Team    string `json:"team"`
+}
+
+type State struct {
+	ID      string
+	Version int64
+	Dir     string
+	Path    string
+}
