--- internal/exploit/output.go
@@ -1,7 +1,0 @@
-package exploit
-
-type Output struct {
-	Name string
-	Team string
-	Out  []byte
-}
