+++ internal/exploit/queue.go
@@ -71,10 +71,10 @@ func (eq *Queue) runExploit(ctx context.Context, et Task) ([]byte, error) {
 	case <-eq.done:
 		cancel()
 		<-errC
-		return res.Bytes(), context.Canceled
+		return res.Bytes(), nil
 	case <-ctx.Done():
 		<-errC
-		return res.Bytes(), context.Canceled
+		return res.Bytes(), ctx.Err()
 	case err := <-errC:
 		return res.Bytes(), err
 	}
