--- internal/exploit/result.go
@@ -1,7 +1,0 @@
-package exploit
-
-type Result struct {
-	Flag    string `json:"flag"`
-	Exploit string `json:"sploit"`
-	Team    string `json:"team"`
-}
