--- internal/exploit/state.go
@@ -1,8 +1,0 @@
-package exploit
-
-type State struct {
-	ID      string
-	Version int64
-	Dir     string
-	Path    string
-}
